requirements.txt
pandas==1.5.0
peewee==3.15.3
psycopg2-binary==2.9.4
PyYAML==6.0